refacto(Map): Move markers to a folder
