+++ frontend/src/components/Map/Markers/ActiveTrekIcon.ts
