+++ frontend/src/components/Map/Markers/TrekIcon.ts
